Extracted SortBottomSheet to generic component
