+++ app/src/main/java/com/projects/moviemanager/MainAppView.kt
@@ -20,14 +20,13 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.projects.moviemanager.common.ui.MainViewModel
-import com.projects.moviemanager.common.ui.components.ModalComponents
+import com.projects.moviemanager.common.ui.components.bottomsheet.ModalComponents
 import com.projects.moviemanager.features.details.DetailsScreen
 import com.projects.moviemanager.navigation.MainNavGraph
 import com.projects.moviemanager.navigation.components.MainNavBar
 import com.projects.moviemanager.navigation.components.MainNavBarItem
 import com.projects.moviemanager.navigation.components.TopNavBar
-import com.projects.moviemanager.common.theme.MovieManagerTheme
-import com.projects.moviemanager.features.watchlist.WatchlistScreen
+import com.projects.moviemanager.common.ui.theme.MovieManagerTheme
 
 @Composable
 fun MainApp() {
