+++ app/src/main/java/com/projects/moviemanager/common/ui/MainViewModel.kt
@@ -3,7 +3,7 @@ package com.projects.moviemanager.common.ui
 import androidx.lifecycle.ViewModel
 import com.projects.moviemanager.common.domain.MediaType
 import dagger.hilt.android.lifecycle.HiltViewModel
-import com.projects.moviemanager.common.ui.components.SortTypeItem
+import com.projects.moviemanager.common.domain.SortTypeItem
 import com.projects.moviemanager.features.home.HomeScreen
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
