+++ app/src/main/java/com/projects/moviemanager/common/ui/components/tab/GenericTabComponents.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
-import com.projects.moviemanager.common.theme.MainBarGreyColor
+import com.projects.moviemanager.common.ui.theme.MainBarGreyColor
 import com.projects.moviemanager.common.ui.util.UiConstants
 import com.projects.moviemanager.util.removeParentPadding
