+++ app/src/main/java/com/projects/moviemanager/features/watchlist/ui/components/WatchlistBottomSheet.kt
@@ -1,32 +1,19 @@
 package com.projects.moviemanager.features.watchlist.ui.components
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Divider
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
 import com.projects.moviemanager.R
 import com.projects.moviemanager.common.domain.MediaType
-import com.projects.moviemanager.common.theme.MainBarGreyColor
 import com.projects.moviemanager.common.ui.MainViewModel
-import com.projects.moviemanager.common.ui.components.SystemNavBarSpacer
-import com.projects.moviemanager.common.ui.components.button.SortButton
-import com.projects.moviemanager.common.ui.util.UiConstants
+import com.projects.moviemanager.common.ui.components.bottomsheet.GenericBottomSheet
+import com.projects.moviemanager.common.ui.components.bottomsheet.SortButton
 import com.projects.moviemanager.util.Constants.UNSELECTED_OPTION_INDEX
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -45,7 +32,8 @@ fun WatchlistSortBottomSheet(
         displaySortScreen(false)
     }
 
-    val onSortButtonClick: (Boolean, Int, WatchlistSortTypeItem) -> Unit = { isButtonSelected, index, sortItem ->
+    val onSortButtonClick: (Boolean, Int, WatchlistSortTypeItem) -> Unit = {
+            isButtonSelected, index, sortItem ->
         selectedIndex = if (isButtonSelected) {
             UNSELECTED_OPTION_INDEX
         } else {
@@ -60,25 +48,9 @@ fun WatchlistSortBottomSheet(
 
     BackHandler { dismissBottomSheet() }
 
-    val textColor = MaterialTheme.colorScheme.onPrimary
-
-    ModalBottomSheet(
-        onDismissRequest = { dismissBottomSheet() },
-        containerColor = MainBarGreyColor
+    GenericBottomSheet(
+        dismissBottomSheet = dismissBottomSheet
     ) {
-        Text(
-            modifier = Modifier
-                .align(Alignment.CenterHorizontally)
-                .offset(y = (-UiConstants.SMALL_MARGIN).dp),
-            text = stringResource(id = R.string.sort_by_header),
-            style = MaterialTheme.typography.headlineMedium,
-            fontWeight = FontWeight.SemiBold,
-            color = textColor
-        )
-        Divider(
-            color = MaterialTheme.colorScheme.inverseSurface,
-            modifier = Modifier.padding(top = UiConstants.SMALL_PADDING.dp)
-        )
         sortOptions.forEachIndexed { index, sortItem ->
             val isButtonSelected = index == selectedIndex
             SortButton(
@@ -86,15 +58,13 @@ fun WatchlistSortBottomSheet(
                     id = R.string.watchlist_sort_button,
                     stringResource(id = sortItem.titleRes)
                 ),
-                textColor = textColor,
+                textColor = MaterialTheme.colorScheme.onPrimary,
                 isSelected = isButtonSelected,
                 onClick = {
                     onSortButtonClick(isButtonSelected, index, sortItem)
                 }
             )
         }
-        Spacer(modifier = Modifier.height(UiConstants.SMALL_PADDING.dp))
-        SystemNavBarSpacer()
     }
 }
