+++ app/src/main/java/com/projects/moviemanager/navigation/components/MainNavBar.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.projects.moviemanager.common.ui.util.UiConstants.BUTTON_NAVIGATION_BAR_HEIGHT
-import com.projects.moviemanager.common.theme.MainBarGreyColor
+import com.projects.moviemanager.common.ui.theme.MainBarGreyColor
 import com.projects.moviemanager.common.ui.MainViewModel
 
 @Composable
